requirements.txt
pandas==2.3.2
numpy==2.3.3
matplotlib==3.10.6
seaborn==0.13.2
scikit-learn==1.7.2
scipy==1.16.1
statsmodels==0.14.5
loguru==0.7.3
langchain==0.3.27
langchain-openai==0.3.33
langchain-groq==0.3.8
uv==0.8.17
python-dotenv==1.1.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
faker==30.8.2
langgraph==0.2.48